Fixes model queries using payableId to use providableId
